requirements.txt
icmplib==3.0.4
PyYAML==6.0.1
